% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_update.R
\name{inventory_update}
\alias{inventory_update}
\title{inventory update}
\usage{
inventory_update(
  inventoryFile,
  removalRecord,
  removeVials = TRUE,
  columnName = "barcode",
  writeToFile = TRUE
)
}
\arguments{
\item{inventoryFile}{character; character string of the .csv file that includes alphanumeric barcodes and any additional vial information of inventory to be updated}

\item{removalRecord}{character; character string of the .csv file that includes alphanumeric barcodes of vials to remove from inventory file}

\item{removeVials}{logical; should rows be removed from inventory? Default TRUE. If FALSE, then rows will not be removed. A new column will be added indicating the vials' inventory status.}

\item{columnName}{character; name of the column that contains the alphanumeric codes. inventoryFile and removalRecord must both have a column with this name containing the alplhanumeric codes. Default "barcode".}

\item{writeToFile}{logical; should updated inventory data be written to disk? Default TRUE. If FALSE, output will not be written to file.}
}
\value{
An updated inventory file in .csv format.
}
\description{
Takes in the unedited vial inventory and outputs the updated inventory as a .csv saved to disk.
}
